+++ GodotSharpDI.Abstractions/ProvideAttribute.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace GodotSharpDI.Abstractions;
 
-[AttributeUsage(AttributeTargets.Property | AttributeTargets.Method, AllowMultiple = false)]
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Method, AllowMultiple = false)]
 public sealed class ProvideAttribute : Attribute
 {
     public Type[] ExposedTypes { get; set; } = [];